:recycle: 미션1 - 팝업 상태 관리 포인트변경 drilling-> 해당 페이지
