+++ src/Routing.jsx
@@ -9,48 +9,37 @@ import useRoute from './hooks/useRoute';
 const Routing = () => {
   const {
     cardInfo,
-    onHandleCardInfoInput,
-    onHandleSave,
-    onHandleSubmit,
+    handleCardInfoInput,
+    handleSave,
+    handleSubmit,
     cardList,
     cardCompanyList,
-    onHandleCompanyPopupClick,
-    isShowPopup,
-    onHandleResetCardInfo,
-    setIsShowPopup
+    handleResetCardInfo,
+    setCardInfo
   } = usePayment();
 
   const { movePage } = useRoute();
 
   return (
     <Routes>
-      <Route
-        path={PATH.HOME}
-        element={<CardList cardList={cardList} movePage={movePage} />}
-      />
+      <Route path={PATH.HOME} element={<CardList cardList={cardList} movePage={movePage} />} />
       <Route
         path={PATH.REGIST}
         element={
           <CardRegist
             cardInfo={cardInfo}
-            onChange={onHandleCardInfoInput}
-            onSubmit={onHandleSubmit}
+            onChange={handleCardInfoInput}
+            onSubmit={handleSubmit}
             cardCompanyList={cardCompanyList}
-            onClickPopup={onHandleCompanyPopupClick}
-            isShowPopup={isShowPopup}
-            onReset={onHandleResetCardInfo}
-            onShow={setIsShowPopup}
+            setCardInfo={setCardInfo}
+            onReset={handleResetCardInfo}
           />
         }
       />
       <Route
         path={PATH.SAVE}
         element={
-          <CardSave
-            onSave={onHandleSave}
-            cardInfo={cardInfo}
-            onChange={onHandleCardInfoInput}
-          />
+          <CardSave onSave={handleSave} cardInfo={cardInfo} onChange={handleCardInfoInput} />
         }
       />
     </Routes>
